+++ test/unit/braintree/webhook_notification_gateway_spec.js
@@ -1037,6 +1037,33 @@ describe("WebhookNotificationGateway", function () {
       );
     });
 
+    it("builds a sample notification for a billing skipped subscription", function (done) {
+      let notification =
+        specHelper.defaultGateway.webhookTesting.sampleNotification(
+          WebhookNotification.Kind.SubscriptionBillingSkipped,
+          "my_id"
+        );
+      let bt_signature = notification.bt_signature;
+      let bt_payload = notification.bt_payload;
+
+      specHelper.defaultGateway.webhookNotification.parse(
+        bt_signature,
+        bt_payload,
+        function (err, webhookNotification) {
+          assert.equal(
+            webhookNotification.kind,
+            WebhookNotification.Kind.SubscriptionBillingSkipped
+          );
+          assert.equal(webhookNotification.subscription.id, "my_id");
+          assert.equal(webhookNotification.subscription.transactions.length, 0);
+          assert.equal(webhookNotification.subscription.discounts.length, 0);
+          assert.equal(webhookNotification.subscription.addOns.length, 0);
+
+          done();
+        }
+      );
+    });
+
     it("builds a sample notification for a successfully charged subscription", function (done) {
       let notification =
         specHelper.defaultGateway.webhookTesting.sampleNotification(